fix : datastore에 저장하는 challenge seq를  long으로 변경

+++ RunWithMe/data/src/main/java/com/side/data/datasource/datastore/DataStoreDataSource.kt
@@ -20,7 +20,7 @@ interface DataStoreDataSource {
 
     fun getRefreshToken(): Flow<String>
 
-    suspend fun saveRunningChallengSeq(challengSeq: Int)
+    suspend fun saveRunningChallengSeq(challengSeq: Long)
 
     suspend fun saveRunningGoalAmount(goalAmount: Long)
 
@@ -28,7 +28,7 @@ interface DataStoreDataSource {
 
     suspend fun saveRunningChallengeName(challengeName: String)
 
-    fun getRunningChallengeSeq(): Flow<Int>
+    fun getRunningChallengeSeq(): Flow<Long>
 
     fun getRunningGoalAmount(): Flow<Long>
